+++ src/main/java/com/bagile/tms/entities/TmsZone.java
@@ -12,18 +12,18 @@ public class TmsZone   extends EmsEntity{
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seq")
     @SequenceGenerator(name = "seq", sequenceName = "seq_tms_zone_id", allocationSize = 1)
     @Column(name = "tms_zoneid", unique = true, nullable = false, precision = 10, scale = 0)
-    private int tmsZoneId;
+    private Long tmsZoneId;
     @Column(name = "tms_zonename")
     private String tmsZoneName;
 
     public TmsZone() {
     }
 
-    public int getTmsZoneId() {
+    public Long getTmsZoneId() {
         return tmsZoneId;
     }
 
-    public void setTmsZoneId(int tmsZoneId) {
+    public void setTmsZoneId(Long tmsZoneId) {
         this.tmsZoneId = tmsZoneId;
     }
